Fix bugs with the navigator
