+++ buildsystem-core/src/main/java/com/eintosti/buildsystem/manager/ArmorStandManager.java
@@ -9,8 +9,8 @@
 package com.eintosti.buildsystem.manager;
 
 import com.cryptomorin.xseries.XMaterial;
+import com.eintosti.buildsystem.object.navigator.NavigatorInventoryType;
 import com.eintosti.buildsystem.util.external.ItemSkulls;
-import com.google.common.collect.Sets;
 import org.bukkit.Location;
 import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Player;
@@ -19,7 +19,6 @@ import org.bukkit.inventory.meta.SkullMeta;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Set;
 import java.util.UUID;
 
 /**
@@ -49,12 +48,12 @@ public class ArmorStandManager {
     }
 
     @SuppressWarnings("deprecation")
-    private ArmorStand spawnArmorStand(Player player, Location location, String customName, boolean customSkull, String skullUrl) {
+    private ArmorStand spawnArmorStand(Player player, Location location, NavigatorInventoryType inventoryType, boolean customSkull, String skullUrl) {
         location.setY(location.getY() - 0.1);
 
         ArmorStand armorStand = location.getWorld().spawn(location, ArmorStand.class);
         armorStand.setVisible(false);
-        armorStand.setCustomName(player.getName() + " × " + customName);
+        armorStand.setCustomName(player.getName() + " × " + inventoryType.getArmorStandName());
         armorStand.setCustomNameVisible(false);
         armorStand.setGravity(false);
         armorStand.setCanPickupItems(false);
@@ -75,17 +74,17 @@ public class ArmorStandManager {
 
     private ArmorStand spawnWorldNavigator(Player player) {
         Location navigatorLocation = calculatePosition(player, SPREAD / 2 * -1);
-        return spawnArmorStand(player, navigatorLocation, "§aWorld Navigator", true, "https://textures.minecraft.net/texture/d5c6dc2bbf51c36cfc7714585a6a5683ef2b14d47d8ff714654a893f5da622");
+        return spawnArmorStand(player, navigatorLocation, NavigatorInventoryType.NAVIGATOR, true, "https://textures.minecraft.net/texture/d5c6dc2bbf51c36cfc7714585a6a5683ef2b14d47d8ff714654a893f5da622");
     }
 
     private ArmorStand spawnWorldArchive(Player player) {
         Location archiveLocation = calculatePosition(player, 0);
-        return spawnArmorStand(player, archiveLocation, "§6World Archive", true, "https://textures.minecraft.net/texture/7f6bf958abd78295eed6ffc293b1aa59526e80f54976829ea068337c2f5e8");
+        return spawnArmorStand(player, archiveLocation, NavigatorInventoryType.ARCHIVE, true, "https://textures.minecraft.net/texture/7f6bf958abd78295eed6ffc293b1aa59526e80f54976829ea068337c2f5e8");
     }
 
     private ArmorStand spawnPrivateWorlds(Player player) {
         Location privateLocation = calculatePosition(player, SPREAD / 2);
-        return spawnArmorStand(player, privateLocation, "§bPrivate Worlds", false, player.getName());
+        return spawnArmorStand(player, privateLocation, NavigatorInventoryType.PRIVATE, false, player.getName());
     }
 
     public void spawnArmorStands(Player player) {
@@ -103,20 +102,14 @@ public class ArmorStandManager {
         }
 
         String playerName = player.getName();
-        Set<String> names = Sets.newHashSet(
-                "§aWorld Navigator",
-                "§6World Archive",
-                "§bPrivate Worlds"
-        );
-
         for (ArmorStand armorStand : armorStands) {
             String customName = armorStand.getCustomName();
             if (customName == null) {
                 continue;
             }
 
-            for (String name : names) {
-                if (customName.equals(playerName + " × " + name)) {
+            for (NavigatorInventoryType inventoryType : NavigatorInventoryType.values()) {
+                if (customName.equals(playerName + " × " + inventoryType.getArmorStandName())) {
                     armorStand.remove();
                     break;
                 }
